Add files via upload
Title: Grid Explorer - Intelligent Robot Design and Programming (CS270) at KAIST

Project Description:
This project aims to implement the Grid Explorer, an intelligent robot designed to detect obstacles and red regions on a 4x6 map. The Grid Explorer's task involves navigating the map, starting from a predetermined point, while having the flexibility to choose the initial direction. It is not necessary for the robot to pass through all the cells on the map.

The main components of the Grid Explorer consist of an EV3 robot, and the primary development tool utilized is Eclipse. The programming language employed for this project is Java.

The path search algorithm implemented in this project is the Depth-First Search (DFS) algorithm, enabling the Grid Explorer to efficiently explore the map and detect obstacles and red regions.